+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## v38.x / v39 (In progress)
+- PC sprites: Added Dog (Laika).
+
 ## v38 (2019-05-10)
 - Ancient Cave: Starstorm grant from Master Barf fight works like Auto-StarMaster. (If fought after using an Auto-Starmaster, fight will grant Omega; if use a Auto-Starmaster after fight, Auto-Starmaster will grant Omega.)
 - Credits: EBSS AC 2019 winner Aurilliux added.